% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_cor_div_habvol}
\alias{check_cor_div_habvol}
\title{Check correlation between alpha diversity and habitat volume}
\usage{
check_cor_div_habvol(in_allvars_sites, alpha_var = "richness")
}
\arguments{
\item{in_allvars_sites}{A list containing a data.table of all merged variables.}

\item{alpha_var}{Character string: name of the alpha diversity column to use (e.g., "richness", "shannon").}
}
\value{
A list of ggplot objects showing the relationships for different organism groups.
}
\description{
Creates diagnostic scatter plots to examine the relationship
between alpha diversity and local habitat volume metrics (e.g., depth, width, velocity),
faceted by country.
}
