% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compile_hydrocon_sites_country}
\alias{compile_hydrocon_sites_country}
\title{Compile hydrological and connectivity data for a country}
\usage{
compile_hydrocon_sites_country(
  in_hydrostats_sub_comb,
  in_STcon_directed,
  in_STcon_undirected,
  in_Fdist_directed,
  in_Fdist_undirected,
  in_site_snapped_gpkg_list,
  in_country
)
}
\arguments{
\item{in_hydrostats_sub_comb}{A list containing hydrological statistics for sites and river network segments.}

\item{in_STcon_directed}{A list of directed STcon data, formatted as an output of \code{postprocess_STcon}.}

\item{in_STcon_undirected}{A list of undirected STcon data.}

\item{in_Fdist_directed}{A list of directed Fdist data.}

\item{in_Fdist_undirected}{A list of undirected Fdist data.}

\item{in_site_snapped_gpkg_list}{A list of paths to snapped site geopackage files.}

\item{in_country}{A character string specifying the country to compile data for.}
}
\value{
A data.table containing all merged hydrological and connectivity metrics.
}
\description{
This function merges various hydrological and connectivity metrics
(hydrological stats, STcon, and Fdist) for a specific country's sites.
}
