% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_cor_matrix}
\alias{compute_cor_matrix}
\title{Compute grouped correlation matrices}
\usage{
compute_cor_matrix(in_allvars_sites)
}
\arguments{
\item{in_allvars_sites}{A list containing the data table (\code{dt}) and a list
of column names (\code{cols}) categorized by their origin.}
}
\value{
A list containing five correlation matrices for different variable
combinations and groupings, plus the original column list.
}
\description{
This function computes correlation matrices for different sets of
variables (hydrology, environment, diversity) and different groupings
(overall, by organism, by country).
}
