% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_ecdf_lookup}
\alias{compute_ecdf_lookup}
\title{Compute left-continuous ECDF values by group.}
\source{
#https://stats.stackexchange.com/questions/585291/is-there-an-equivalent-to-an-ecdf-with-a-sign
https://math.stackexchange.com/questions/1807120/why-arent-cdfs-left-continuous/1807136#1807136
}
\usage{
compute_ecdf_lookup(in_dt, ecdf_column, grouping_columns, na.rm = TRUE)
}
\arguments{
\item{in_dt}{Input data.table.}

\item{ecdf_column}{Column to compute ECDF for.}

\item{grouping_columns}{Columns to group by.}

\item{na.rm}{Remove NAs in ECDF column.}
}
\value{
Data.table with ECDF values by group.
}
\description{
Compute left-continuous ECDF values by group.
}
\details{
If nine 0s and one 1, then 0s are given 0 and 1 is given 0.9
}
