% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_ecdf_multimerge}
\alias{compute_ecdf_multimerge}
\title{Merge multiple ECDF lookups into one table.}
\usage{
compute_ecdf_multimerge(
  in_dt,
  ecdf_columns,
  grouping_columns,
  keep_column,
  na.rm = TRUE
)
}
\arguments{
\item{in_dt}{Input data.table.}

\item{ecdf_columns}{List of columns to compute ECDF for.}

\item{grouping_columns}{Columns to group by.}

\item{keep_column}{Column to keep in output.}

\item{na.rm}{Logical: remove NAs.}
}
\value{
Merged data.table with ECDF columns.
}
\description{
Merge multiple ECDF lookups into one table.
}
