% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_Fdist_rolling}
\alias{compute_Fdist_rolling}
\title{Compute rolling Fdist}
\usage{
compute_Fdist_rolling(in_Fdist_dt, in_sites_dt)
}
\arguments{
\item{in_Fdist_dt}{A data.table containing Fdist values, with 'date' and 'UID' columns.}

\item{in_sites_dt}{A data.table of site information, used to filter for relevant UIDs.}
}
\value{
A data.table with new columns for the rolling mean and max Fdist values.
}
\description{
This function computes rolling mean and maximum values of Fdist
(distance to nearest wet site) for a set of sites.
}
