% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_hydrostats_drn}
\alias{compute_hydrostats_drn}
\title{Compute hydrological statistics for a DRN}
\usage{
compute_hydrostats_drn(
  in_network_path,
  in_sites_dt,
  varname,
  in_hydromod_filled,
  in_network_idcol = "cat"
)
}
\arguments{
\item{in_network_path}{A character string specifying the path to the input network shapefile.}

\item{in_sites_dt}{A data.table containing site information, including the \code{reach_id} for each site.}

\item{varname}{A character string specifying the variable to compute statistics for.
Can be 'isflowing' for intermittence or 'qsim' for discharge.}

\item{in_network_idcol}{A character string specifying the name of the column
in the network that serves as a unique identifier for reaches. Defaults to 'cat'.}

\item{in_hydromod_drn}{A list containing hydrological model data,
typically including a data.table of time series data (\code{data_all}) and a
data.table with date formats (\code{dates_format}).}
}
\value{
A list or data.table containing the computed hydrological statistics.
}
\description{
This function computes hydrological statistics of intermittence
and discharge for a given DRN.
It can compute statistics at the drying river network (DRN) scale or at the
scale of individual sites.
}
