% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_hydrostats_intermittence}
\alias{compute_hydrostats_intermittence}
\title{Compute intermittence statistics for hydrological data.}
\usage{
compute_hydrostats_intermittence(in_hydromod_dt, in_sites_dt, scale = "all")
}
\arguments{
\item{in_hydromod_dt}{Hydrological time series data.table.}

\item{in_sites_dt}{Sites data.table.}

\item{scale}{Scale for computation: 'all', 'drn', or 'site'.}
}
\value{
List of statistics by scale.
}
\description{
Compute intermittence statistics for hydrological data.
}
