% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaTemp_function_M_edit.R
\name{compute_stcon}
\alias{compute_stcon}
\title{Calculate Spatiotemporal Connectivity Index}
\usage{
compute_stcon(
  sites_status_matrix,
  network_structure,
  direction,
  routing_mode = "in",
  weighting = FALSE,
  dist_matrix,
  weighting_links = FALSE,
  link_weights,
  indirect_dispersal = TRUE,
  standardize_neighbors = FALSE,
  value_s_link = 1L,
  value_t_link = 1L,
  value_no_s_link = 0L,
  value_no_t_link = 0L,
  legacy_effect = 1L,
  legacy_length = 1L,
  convert_to_integer = TRUE,
  rounding_factor = 100L,
  verbose = T,
  output = c("Main_matrix", "STconmat", "STcon")
)
}
\arguments{
\item{sites_status_matrix}{A matrix representing the status of the site (wet/dry,
active/inactive).
The dataset should have columns for each site and rows for each monitored day.
Warning: no other columns should be included (e.g., date, other IDs)}

\item{network_structure}{A square matrix representing the basic connections among
sites (adjacency matrix): for a given site (row), each adjacent connected
site (column) is given a value of 1, all others 0. Must have the same number
of rows and columns as there are columns in sites_status_matrix.}

\item{direction}{The direction of the graph, either "directed" or "undirected".}

\item{routing_mode}{The direction for graph connectivity when directed,
can be "in" (routing from upstream if directed), "out" (routing from
downstream if directed), or "all". See ?igraph or ?igraph::closeness
for a better understanding.}

\item{weighting}{Logical; whether to weight the connectivity based on distances.}

\item{dist_matrix}{A distance matrix representing the distances between sites.
Can be any type of distance (euclidean, environmental, topographic, ...)
between pairs of sites}

\item{weighting_links}{Logical; whether to use specific data for each time
unit to quantify connections.}

\item{link_weights}{A list of data frames with the same structure as
\code{sites_status_matrix}, providing daily flow data.}

\item{indirect_dispersal}{Logical; whether to consider indirect dispersal
for lost nodes (i.e., a bump in dispersal before disconnection).}

\item{standardize_neighbors}{Logical; whether to standardize STCon based
on the number of possible connections (i.e., number of upstream neighbors
for downstream nodes when routing_mode==in, as downstream nodes will
necessarily have higher values everything being equal).}

\item{value_s_link}{Numeric; value attributed to each spatial link.}

\item{value_t_link}{Numeric; value attributed to each temporal link.}

\item{value_no_s_link}{Numeric; value attributed to spatial disconnections.}

\item{value_no_t_link}{Numeric; value attributed to temporal disconnections.}

\item{legacy_effect}{Numeric; weight given to the link/no-link values for each time unit.
It is a vector of length legacy_length with values ranging from 0 to 1 to
modulate the relevance of a connection through time.}

\item{legacy_length}{Numeric; number of time units considered for legacy effects.
Set to 1 by default (i.e., the connectivity of time T1 will only be considered for T2).
If value is set at legacy_length=3, the connectivity of T1 will be considered
for T2, T3, and T4.}

\item{convert_to_integer}{Logical; whether to convert the output to integers
to minimize storage of outputs.}

\item{rounding_factor}{Numeric; factor used for rounding the output before
converting to integer (can be a fraction of 1 when distances are large so that
STCon < 2*10^9 to be converted).}

\item{verbose}{Logical; whether to print messages during the function execution.}

\item{output}{A character vector specifying the output components:
'Main_matrix', 'STconmat', 'STcon'.}
}
\value{
A list containing the spatiotemporal connectivity matrix,
the collapsing matrix (STconmat), and the standardized connectivity (STcon).
}
\description{
This function calculates spatiotemporal connectivity indices based on
spatiotemporal graph networks. The aim is to provide a methodological framework
for calculating these indices based on high-frequency information obtained
from natural ecosystems.
}
\details{
The system does not necessarily need to quantify water absence/presence
but each cell value must represent a feature defining connectivity "on" or "off"
and that can be transmitted to build ecologically meaningful links in a
spatiotemporal graph. Requires: assertthat, data.table, igraph, magrittr, purrr
}
\references{
The current script is a modification of the original function introduced in:
Cunillera-Montcusi, D., Fernandez-Calero, J. M., Polsterl, S., Valera, J.,
Argelich, R., Cid, N., Bonada, N., & Canedo-Arguelles, M. (2023).
Navigating through space and time: a methodological approach to quantify
spatiotemporal connectivity using flow intermittence data as a study case.
}
