% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_sitepoints_raw}
\alias{create_sitepoints_raw}
\title{Create spatial point features from site data.}
\usage{
create_sitepoints_raw(
  in_dt,
  lon_col,
  lat_col,
  out_points_path,
  columns_to_include = NULL
)
}
\arguments{
\item{in_dt}{Input data.table of sites.}

\item{lon_col}{Longitude column name.}

\item{lat_col}{Latitude column name.}

\item{out_points_path}{Path for output points file.}

\item{columns_to_include}{Optional: columns to include in output.}
}
\value{
Path to output points file.
}
\description{
Create spatial point features from site data.
}
