% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_sites_gpkg}
\alias{create_sites_gpkg}
\title{Create sites GeoPackage}
\usage{
create_sites_gpkg(
  in_hydromod_paths_dt,
  in_sites_dt,
  out_dir,
  geom,
  in_network_path_list = NULL,
  in_network_idcol = "cat",
  overwrite = FALSE
)
}
\arguments{
\item{in_hydromod_paths_dt}{A data.table with a list of paths to hydrological
model input/output data for each country.}

\item{in_sites_dt}{A data.table containing site information.}

\item{out_dir}{A character string specifying the output directory for the GeoPackage files.}

\item{geom}{A character string specifying the geometry type to create, either 'reaches' or 'points'.}

\item{in_network_path_list}{A list of character strings specifying the paths
to network shapefiles for each country. Required if \code{geom} is 'reaches'.}

\item{in_network_idcol}{A character string specifying the ID column for network reaches. Defaults to 'cat'.}

\item{overwrite}{A logical value. If \code{TRUE}, existing files will be overwritten. Defaults to \code{FALSE}.}
}
\value{
A named vector of character strings, where names are countries and
values are the paths to the created GeoPackage files.
}
\description{
This function creates a GeoPackage (\code{.gpkg}) file containing
either site points or river reaches associated with specific countries.
}
