% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_ssn_europe}
\alias{create_ssn_europe}
\title{Create a European-scale SSN}
\usage{
create_ssn_europe(
  in_network_path,
  in_sites_path,
  in_allvars_dt,
  in_local_env_pca,
  in_barriers_path = NULL,
  in_hydrostats_net_hist,
  in_pred_pts = NULL,
  out_dir,
  out_ssn_name,
  overwrite = T
)
}
\arguments{
\item{in_network_path}{A list of file paths to the river network shapefiles.}

\item{in_sites_path}{A list of file paths to the site observation points.}

\item{in_allvars_dt}{A data table containing all observation data,
including diversity and environmental metrics.}

\item{in_local_env_pca}{The output from the \code{ordinate_local_env} function,
containing PCA axes for environmental data.}

\item{in_barriers_path}{A list of file paths to dam and barrier locations.}

\item{in_hydrostats_net_hist}{Historical hydrological statistics for the network.}

\item{in_pred_pts}{An optional list of historical and projected prediction points from \code{create_ssn_preds}.}

\item{out_dir}{The directory to save the output SSN files.}

\item{out_ssn_name}{The base name for the output SSN file.}

\item{overwrite}{A logical value indicating whether to overwrite existing files.}
}
\value{
A list of SSN objects, one for each organism.
}
\description{
This function assembles a full Spatial Stream Network (SSN) for
Europe by integrating the river network with observation sites, barriers,
and prediction points. It calculates key upstream metrics to support spatial modeling.
}
