% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_ssn_pred_pts}
\alias{create_ssn_pred_pts}
\title{Create Spatial Stream Network prediction points}
\usage{
create_ssn_pred_pts(in_network_path)
}
\arguments{
\item{in_network_path}{A list of file paths to the river network shapefiles, separated by country.}
}
\value{
A list containing two spatial objects: one for historical prediction
points and one for projected prediction points.
}
\description{
This function prepares spatial data for SSN modeling by creating
prediction points from a river network and associating them with historical
and projected hydrological data.
}
