% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{define_hydromod_paths}
\alias{define_hydromod_paths}
\title{Define hydrological model input and output data paths for the 6 DRNs}
\usage{
define_hydromod_paths(in_hydromod_dir)
}
\arguments{
\item{in_hydromod_dir}{Character. Path to the root hydrological model directory.}
}
\value{
A \code{data.table} with country, catchment, best simulation number,
and standardized file paths for each catchment.
}
\description{
Creates a table of file paths related to hydrological
simulations for several European catchments. It standardizes the paths
to key files: NetCDF outputs, shapefiles, parameter files, and reach IDs.
}
