% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{direct_network}
\alias{direct_network}
\title{Direct entire river network}
\usage{
direct_network(rivnet_path, idcol, outletid, outdir = NULL, save_gpkg = FALSE)
}
\arguments{
\item{net}{\code{sf} object of river segments.}

\item{outlet_id}{Integer; ID of the outlet segment (\code{UID}).}
}
\value{
\code{sf} object with all segments oriented downstream.
}
\description{
Orients all river segments in a network to flow downstream, beginning from a
specified outlet segment. The function assumes the network has a single outlet
(a dangling segment at the river mouth).
}
