% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{direct_network_inner}
\alias{direct_network_inner}
\title{Recursively ensures river segments are oriented downstream, starting from a
specified segment (typically the outlet). If a segment is reversed relative
to its downstream neighbor, its geometry is reversed.}
\usage{
direct_network_inner(segment, in_network, idcol, visited = NULL)
}
\arguments{
\item{net}{\code{sf} object of river segments.}

\item{segid}{Integer; ID of the current segment to check.}

\item{check_ids}{Integer vector of segment IDs already checked (used to
avoid infinite recursion).}
}
\value{
\code{sf} object with geometries reversed as needed.
}
\description{
Recursively ensures river segments are oriented downstream, starting from a
specified segment (typically the outlet). If a segment is reversed relative
to its downstream neighbor, its geometry is reversed.
}
\keyword{internal}
