% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dist_proj}
\alias{dist_proj}
\title{Create a two-point equidistant projection string for a bounding box}
\usage{
dist_proj(x)
}
\arguments{
\item{x}{An \code{sf} object (or anything accepted by \code{sf::st_bbox}).}
}
\value{
A PROJ.4 projection string for a two-point equidistant projection
with WGS84 ellipsoid and units in meters.
}
\description{
Defines a standard two-point equidistant (tpeqd) projection based on the
bounding box of a spatial object. Useful for distance calculations that
require a locally suitable projection.
}
\references{
Based on: \url{https://gis.stackexchange.com/questions/313721/automatically-get-an-adequate-projection-system-based-on-a-bounding-box}
}
