% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_unzip}
\alias{download_unzip}
\title{Download and unzip a file from a URL to a specified directory.}
\usage{
download_unzip(url, out_dir, download_mode = "wb", out_zip = NULL)
}
\arguments{
\item{url}{URL of the file to download.}

\item{out_dir}{Output directory where the file will be unzipped.}

\item{download_mode}{Mode for download.file (default 'wb' for binary).}

\item{out_zip}{Optional: Path for the downloaded zip file.}
}
\value{
Output directory.
}
\description{
Download and unzip a file from a URL to a specified directory.
}
