% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fix_complex_confluences}
\alias{fix_complex_confluences}
\title{Fix complex confluences in a river network}
\usage{
fix_complex_confluences(
  rivnet_path,
  max_node_shift = 5,
  outdir = NULL,
  out_path = NULL
)
}
\arguments{
\item{rivnet_path}{A character string specifying the path to the input river
network shapefile. The network must be directed and topologically correct
except for the complex confluences.}

\item{max_node_shift}{A numeric value specifying the maximum distance to
shift nodes in order to fix confluences. Defaults to 5.}

\item{outdir}{A character string specifying the output directory for the
fixed network file. If not specified, the output file will be written to
the same directory as the input.}

\item{out_path}{A character string specifying the full path for the output file.
If provided, \code{outdir} is ignored.}
}
\value{
A character string with the file path to the newly created fixed network shapefile.
}
\description{
This function takes a directed river network shapefile and
fixes complex confluences by shifting nodes to resolve topological errors.
It relies on the \code{fix_confluences_inner} function, a helper function.
}
