% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_covariate_name}
\alias{get_covariate_name}
\title{Identify target covariate}
\usage{
get_covariate_name(fit_mod, hydro_var)
}
\arguments{
\item{fit_mod}{A fitted SSN2 model object}

\item{hydro_var}{The hydrological variable name (may be root name)}
}
\value{
Character string of the exact covariate name in the model
}
\description{
Makes sure that the provided hydrological covariate is in the formula
Otherwise, extracts the exact name, handling cases where the covariate might
have scaling suffixes.
}
\examples{
get_covariate_name(fit_mod, "DurD") # Returns "DurD180past_scaled"

}
