% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_hydrowindow_emmeans}
\alias{get_hydrowindow_emmeans}
\title{Extract and plot marginal means (EMMeans) from best models}
\usage{
get_hydrowindow_emmeans(
  best_dt,
  permutations_dt,
  in_hydro_vars_dt,
  in_drn_dt,
  plot = T
)
}
\arguments{
\item{best_dt}{A \code{data.table} of best models ($best) returned by \code{\link[=prepare_hydrowindow_perf_table]{prepare_hydrowindow_perf_table()}}.}

\item{in_hydro_vars_dt}{Metadata table with hydro variable names and labels.}
}
\value{
A \code{data.table} and plots of estimated marginal means across predictors.
}
\description{
Runs \code{\link[=get_ssn_emmeans]{get_ssn_emmeans()}} for all predictors in the best models.
}
\note{
This approach is adequate even for models with quadratic terms
as long as they are explicitly included in the model with a I() or poly()
}
\references{
\href{https://rvlenth.github.io/emmeans/articles/basics.html#depcovs}{emmeans documentation}
}
