% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_hydrowindow_predictions}
\alias{get_hydrowindow_predictions}
\title{Get predictions from best models}
\usage{
get_hydrowindow_predictions(best_dt)
}
\arguments{
\item{best_dt}{A \code{data.table} ($best) returned by \code{\link[=prepare_hydrowindow_perf_table]{prepare_hydrowindow_perf_table()}}.}
}
\value{
A \code{data.table} of observed and fitted values with hydro variable metadata.
}
\description{
Uses \code{SSN2::augment()} to obtain fitted vs observed values and attaches metadata.
}
