% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_nc_var_present}
\alias{get_nc_var_present}
\title{Read variable from NetCDF and format with hydrological year and season info.}
\usage{
get_nc_var_present(nc, varname, reachID, dates, selected_sims = NULL)
}
\arguments{
\item{nc}{NetCDF object.}

\item{varname}{Character. Variable name to extract.}

\item{reachID}{Vector of reach IDs.}

\item{dates}{Vector of dates.}

\item{selected_sims}{Optional: which simulation layers to select.}
}
\value{
List: formatted data and date info.
}
\description{
Read variable from NetCDF and format with hydrological year and season info.
}
