% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hsaValidHydroYear}
\alias{hsaValidHydroYear}
\title{Assess validity of hydrological years based on completeness and missing values.}
\usage{
hsaValidHydroYear(hy, x = NULL, n = 365, na.th = 0.005)
}
\arguments{
\item{hy}{Vector. Hydrological year assignments (will be coerced to factor)}

\item{x}{Matrix or vector with same length as hy, used to check missing values.}

\item{n}{Integer. Minimum length of a complete hydrological year.}

\item{na.th}{Numeric. Maximum tolerated proportion (threshold) of missing values.}
}
\value{
Logical vector indicating valid time steps.
}
\description{
Assess validity of hydrological years based on completeness and missing values.
}
\details{
Given a hydrological year vector and a matrix with the same number
of rows (and any number of column), this function returns a logical
vector of the same length as the hydrological year vector that
indicates whether or not the time steps are part of a valid
hydrological year. Validity of a hydrological year is assessed
according the two following rules:
\itemize{
\item is the year complete (i.e. is there at least 'n' (365) days)?
\item is there less than 'na.th' (proportion) missing values in
the year?
}
}
