% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{manual_clean_croatia}
\alias{manual_clean_croatia}
\title{Manual cleaning for Croatia river network}
\usage{
manual_clean_croatia(in_net)
}
\arguments{
\item{rivnet}{\code{sf} object of the river network.}
}
\value{
A manually corrected \code{sf} object.
}
\description{
Fixes geometrical issues with specific segments (hard-coded IDs) by replacing
geometries and splitting problem lines.
}
