% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_ssn_facets}
\alias{map_ssn_facets}
\title{Create a faceted SSN map}
\usage{
map_ssn_facets(
  in_ssn,
  facet_col,
  in_pts = NULL,
  linewidth_col = "qsim_avg",
  linecolor_col = NULL,
  shape_col = NULL,
  ptcolor_col = NULL,
  nbreaks = NULL,
  page_title = NULL,
  cover_zero = FALSE
)
}
\arguments{
\item{in_ssn}{A single SSN object.}

\item{facet_col}{The column name to be used for faceting.}

\item{in_pts}{An optional name of the point set to plot (e.g., 'obs').}

\item{linewidth_col}{The column name for controlling river segment line width.}

\item{linecolor_col}{An optional column for coloring river segments.}

\item{shape_col}{An optional column for controlling point shape.}

\item{ptcolor_col}{An optional column for coloring points.}

\item{page_title}{An overall title for the plot.}
}
\value{
A patchwork object containing the faceted maps.
}
\description{
This function generates a multi-panel map of an SSN, with each
panel representing a facet category (e.g., country). It ensures a consistent,
square aspect ratio across all panels for visual comparability.
}
