% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_ssn_summarized}
\alias{map_ssn_summarized}
\title{Map summarized SSN variables}
\usage{
map_ssn_summarized(
  in_ssn_summarized,
  in_allvars_summarized,
  in_organism_dt,
  verbose = T
)
}
\arguments{
\item{in_ssn_summarized}{A list of SSN objects, typically for different organisms.}

\item{in_allvars_summarized}{A list containing a data table of all variables and column names.}

\item{in_organism_dt}{A data table mapping organism codes to their labels.}

\item{verbose}{A logical value to indicate whether to print progress messages.}
}
\value{
A named list of ggplot objects, combining maps of diversity and physical variables.
}
\description{
Generates a series of faceted maps visualizing diversity and
physical variables across the European SSN. It creates separate plots
for each variable, faceted by country.
}
