% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_ssn_util}
\alias{map_ssn_util}
\title{Plot a single SSN map}
\usage{
map_ssn_util(
  in_ssn,
  in_edges,
  linewidth_col = "qsim_avg",
  linecolor_col = NULL,
  linecolor_lims = NULL,
  in_pts = NULL,
  shape_col = NULL,
  ptcolor_col = NULL,
  ptcolor_lims = NULL,
  nbreaks = NULL,
  cover_zero = FALSE
)
}
\arguments{
\item{in_ssn}{A single SSN object.}

\item{in_edges}{The edges (river segments) of the SSN to be plotted.}

\item{linewidth_col}{The name of the column to control river segment line width.}

\item{linecolor_col}{An optional column for coloring river segments.}

\item{linecolor_lims}{Optional limits for the line color scale.}

\item{in_pts}{Optional points (e.g., sites) to be plotted on the network.}

\item{shape_col}{An optional column for controlling point shape.}

\item{ptcolor_col}{An optional column for coloring points.}

\item{ptcolor_lims}{Optional limits for the point color scale.}
}
\value{
A ggplot object representing the SSN map.
}
\description{
This utility function creates a single ggplot map of an SSN,
allowing for visualization of network edges and points.
It supports customization of line width, line color, point color,
and point shape based on specified data columns.
}
