% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{merge_allvars_summarized}
\alias{merge_allvars_summarized}
\title{Merge all variables summarized by site}
\usage{
merge_allvars_summarized(
  in_spdiv_local,
  in_hydrocon_summarized,
  in_env_summarized,
  dry_only_sites,
  dtcols_sites
)
}
\arguments{
\item{in_hydrocon_summarized}{Hydrological/connectivity data summarized by site.}

\item{in_env_summarized}{A list of summarized environmental data.tables.}

\item{dry_only_sites}{Sites that never flowed (from merge_allvars_sites).}

\item{dtcols_sites}{Column lists from merge_allvars_sites.}

\item{spdiv}{A biodiversity data.table (already cleaned).}

\item{in_genal_upa}{Upstream area data for the Genal basin.}
}
\value{
A list containing:
\item{dt_summarized}{Merged site-level summarized data}
\item{cols}{List of relevant column vectors for summarized merging}
}
\description{
Summarizes biodiversity across campaigns and merges with
site-level hydrological & environmental summaries.
}
