% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mergeDTlist}
\alias{mergeDTlist}
\title{Merge a list of data.tables, adding suffixes by source table name.}
\usage{
mergeDTlist(dt_list, by = NULL, all = TRUE, sort = FALSE, set_suffix = TRUE)
}
\arguments{
\item{dt_list}{List of data.tables.}

\item{by}{Columns to merge by.}

\item{all}{Logical: Merge all rows.}

\item{sort}{Logical: Sort merged table.}

\item{set_suffix}{Logical: Add suffix by name.}
}
\value{
Merged data.table.
}
\description{
Merge a list of data.tables, adding suffixes by source table name.
}
