% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{model_fun_biof_richness_yr}
\alias{model_fun_biof_richness_yr}
\title{Fungi model for annual data}
\usage{
model_fun_biof_richness_yr(
  in_ssn_eu_summarized,
  in_allvars_summarized,
  in_cor_matrices,
  ssn_covtypes,
  scale_predictors = T,
  interactive = F
)
}
\arguments{
\item{in_ssn_eu_summarized}{A list containing the summarized SSN object for
macroinvertebrates. Expected to have a structure like \code{in_ssn_eu_summarized$miv_nopools$ssn}.}

\item{in_allvars_summarized}{A list containing merged environmental and hydrological
variables. This is used to define the list of candidate variables.}

\item{in_cor_matrices}{A list of correlation matrices, used to check for
relationships between variables and alpha diversity.}

\item{ssn_covtypes}{A data frame or list specifying SSN covariance types.
This is used to systematically test different spatial covariance structures.}

\item{scale_predictors}{A logical value. If TRUE, all candidate hydrological
predictor variables are scaled to have a mean of 0 and SD of 1. Defaults to TRUE.}

\item{interactive}{A logical value. If TRUE, the function generates a series
of exploratory plots and model diagnostics for interactive inspection. Defaults to FALSE.}
}
\value{
A list containing the following:
\itemize{
\item \strong{model_selection_table}: A data table summarizing the performance
(e.g., AIC, VIF) of all tested models.
\item \strong{ssn_mod_fit}: The final fitted SSN model selected for prediction.
\item \strong{ssn_pred_final}: The augmented data frame with predictions from the
prediction model.
\item \strong{torgegram_mod_pred}: A diagnostic plot (ggplot object) showing the
Torgegram for the prediction model's residuals.
\item \strong{ssn_mod_fit_best}: The "absolute best" model fit, refitted with REML.
\item \strong{ssn_pred_best}: Augmented data for the "best" model.
}
}
\description{
This function models macroinvertebrate alpha diversity averaged across all samplig dates.
It performs data preparation, conducts exploratory data analysis (if \code{interactive} is TRUE),
tests various model structures using a streamlined SSN generalized linear
modeling approach, and provides a summary of model performance.
}
\details{
The function's primary goal is to identify the best-fitting spatial stream
network (SSN) model for predicting macroinvertebrate alpha diversity based on
hydrological, environmental, and spatial variables.
}
