% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ordinate_local_env}
\alias{ordinate_local_env}
\title{Ordinate local environmental data}
\usage{
ordinate_local_env(in_allvars_dt)
}
\arguments{
\item{in_allvars_dt}{A data table containing environmental data, which may be summarized by site.}
}
\value{
A list containing the PCA results for macroinvertebrates and eDNA groups,
as well as a combined data table with the new PCA axes.
}
\description{
This function performs Principal Component Analysis (PCA) on a
dataset of local environmental variables to create a reduced set of
dimensions (PCA axes) that capture the main sources of environmental variation.
}
