% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pad_ssn_map}
\alias{pad_ssn_map}
\title{Pad SSN map to a square bounding box}
\usage{
pad_ssn_map(in_edges = NULL, in_pts = NULL)
}
\arguments{
\item{in_edges}{The river network edges (an \code{sf} object).}

\item{in_pts}{The observation points (an \code{sf} object).}
}
\value{
A list with \code{xlim} and \code{ylim} vectors for the new square plot limits.
}
\description{
A utility function that calculates a square bounding box for an SSN map. This ensures consistent aspect ratios across multiple faceted plots, even if the geographic extents of the data differ.
}
