% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_cor_heatmaps}
\alias{plot_cor_heatmaps}
\title{Plot correlation heatmaps}
\usage{
plot_cor_heatmaps(in_cor_matrices, p_threshold = 1)
}
\arguments{
\item{in_cor_matrices}{A list of correlation matrices, typically the output
of \code{compute_cor_matrix} or \code{compute_cor_matrix_summarized}.}

\item{p_threshold}{A numeric value to filter correlations based on their p-value.}
}
\value{
ggplot of heatmap
}
\description{
This function generates a set of heatmaps to visualize the
correlation matrices computed previously. It can display correlations for
different variable sets and across various groupings like organism and country.
}
