% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_edna_biof_vs_sedi}
\alias{plot_edna_biof_vs_sedi}
\title{Plot eDNA biofilm vs sediment}
\usage{
plot_edna_biof_vs_sedi(in_allvars_sites)
}
\arguments{
\item{in_allvars_sites}{A list containing the merged data tables, specifically \code{in_allvars_sites$dt}.}
}
\value{
A list containing two ggplot objects: \code{richness} and \code{site_gamma}.
}
\description{
This function creates two plots comparing richness and Gamma
diversity between eDNA samples from biofilm and sediment, faceted by organism
class and country.
}
