% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_hydro_comparison}
\alias{plot_hydro_comparison}
\title{Compare standard vs. flow-duration curve-based hydro metrics}
\usage{
plot_hydro_comparison(var_substr, in_cor_dt, color_list)
}
\arguments{
\item{in_cor_dt}{A data.table containing correlation matrices, including \code{correlation} and \code{p_value}.}

\item{color_list}{A named vector of colors for plotting each country.}
}
\value{
A ggplot object of the comparison plot.
}
\description{
This function generates a scatter plot to compare the correlation
coefficients of two different hydrological metrics against a biological
response variable. It helps to visualize how closely two different measures
of flow, such as \code{DurD} (duration) and \code{PDurD} (proportion of duration),
relate to ecological outcomes.
}
