% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_hydrowindow_obs_preds}
\alias{plot_hydrowindow_obs_preds}
\title{Plot observed vs predicted values}
\usage{
plot_hydrowindow_obs_preds(preds, resp_var)
}
\arguments{
\item{preds}{A \code{data.table} returned by \code{\link[=get_predictions]{get_predictions()}}.}

\item{resp_var}{Character. Name of the response variable.}
}
\value{
A \code{ggplot2} object.
}
\description{
Scatterplot of observed vs fitted values for each hydro variable and country.
}
