% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_hydrowindow_x_preds}
\alias{plot_hydrowindow_x_preds}
\title{Plot predictors vs fitted values}
\usage{
plot_hydrowindow_x_preds(preds, best_dt)
}
\arguments{
\item{preds}{A \code{data.table} returned by \code{\link[=get_predictions]{get_predictions()}}.}

\item{best_dt}{A \code{data.table} ($best) returned by \code{\link[=prepare_hydrowindow_perf_table]{prepare_hydrowindow_perf_table()}}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Scatterplots of predictor variables against fitted values across countries.
}
