% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_ssn_mod_diagplot}
\alias{plot_ssn_mod_diagplot}
\title{Get diagnostic plots of SSN Model Performance}
\usage{
plot_ssn_mod_diagplot(
  in_mod_fit,
  in_perf_dt,
  in_drn_dt,
  in_organism_dt,
  in_hydro_vars_dt,
  write_plots = T,
  response_var_label,
  plot_path_prefix = NULL,
  out_dir
)
}
\arguments{
\item{write_plots}{A logical value. If TRUE, the generated plots are saved to
the directory specified by \code{out_dir}.}

\item{out_dir}{A character string specifying the output directory for plots
if \code{write_plots} is TRUE.}

\item{in_ssn_mods}{A list of fitted SSN models, typically generated by another
function (e.g., \code{model_miv_yr}). It is expected to contain \code{ssn_pred_best},
\code{ssn_pred_final}, \code{ssn_mod_fit_best}, and \code{ssn_mod_fit} objects.}
}
\value{
A list containing the variance components for the "best" and "final" models.
\itemize{
\item \strong{varcomp_best}: Variance components for the best-fit model.
\item \strong{varcomp_final}: Variance components for the final prediction model.
}
}
\description{
This function takes a list of fitted SSN models and performs diagnostics. It creates
plots of predicted vs. observed values for both the "best" and "final" models
and optionally saves them to a specified directory. It also returns variance
components for both models.
}
