% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_ssn_proj}
\alias{plot_ssn_proj}
\title{Map SSN model projections and changes}
\usage{
plot_ssn_proj(
  in_future_sims_dt,
  in_future_stats_dt,
  in_drn_dt,
  in_env_summarized,
  in_mod_fit_list,
  in_organism_dt,
  mod_sub = c("miv_richness", "ept_richness", "och_richness", "dia_sedi_richness",
    "fun_sedi_richness"),
  write_plots = T,
  out_dir = NULL
)
}
\arguments{
\item{out_dir}{A character string specifying the output directory where the
generated maps and plots will be saved.}

\item{in_ssn}{The input SSN object.}

\item{in_ssn_mods}{A list of fitted SSN models, as returned by \code{model_miv_yr}.}

\item{in_ssn_preds}{A list of historical and projected predictions, as returned by
\code{predict_ssn_mod}.}
}
\value{
The function is primarily used for its side effects (creating plots and
maps) and does not return a value.
}
\description{
Map SSN model projections and changes
}
