% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_STcon}
\alias{plot_STcon}
\title{Plot STcon values on a river network}
\usage{
plot_STcon(
  in_STcon_list,
  in_date,
  in_window = 10,
  in_net_shp_path,
  reverse_weighted_stcon = TRUE
)
}
\arguments{
\item{in_STcon_list}{A list containing a data.table of STcon values, typically the output of \code{postprocess_STcon}.}

\item{in_date}{A Date object specifying the date to plot.}

\item{in_window}{An integer specifying the window size to plot. Defaults to \code{10}.}

\item{in_net_shp_path}{A character string specifying the path to the network shapefile.}

\item{reverse_weighted_stcon}{A logical value. If \code{TRUE},
the STcon values are inverted for plotting (e.g., lower STcon corresponds to a darker color).
Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot} object representing the plot.
}
\description{
This function takes a formatted list of STcon results,
a specific date, and a window size, and generates a \code{ggplot} object to
visualize the STcon values on a river network.
}
