% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{postprocess_STcon}
\alias{postprocess_STcon}
\title{Post-process STcon results}
\usage{
postprocess_STcon(
  in_STcon,
  in_net_shp_path,
  standardize_STcon = FALSE,
  in_STcon_ref = NULL
)
}
\arguments{
\item{in_STcon}{A list of STcon results, typically the output of \code{compute_STcon_rolling}.}

\item{in_net_shp_path}{A character string specifying the path to the network shapefile.}

\item{standardize_STcon}{A logical value. If \code{TRUE}, STcon values are
standardized by a reference dataset. Defaults to \code{FALSE}.}

\item{in_STcon_ref}{A list of reference STcon results for standardization.
Required if \code{standardize_STcon} is \code{TRUE}.}
}
\value{
A list containing a long-format data table of STcon values (\code{STcon_dt})
and a list of STcon matrices (\code{STcon_mat}).
}
\description{
This function takes the raw output of STcon computation,
standardizes the values, and formats the results into a long data table
and a list of matrices for easier use.
}
