% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict_ssn_mod}
\alias{predict_ssn_mod}
\title{Predict future richness from a fitted SSN model}
\usage{
predict_ssn_mod(
  in_ssn_mod_fit,
  in_hydrocon_sites_proj,
  type_predict,
  predict_years,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{in_ssn_mods}{A list of fitted SSN models, typically generated by \code{model_miv_yr}.
It is expected to contain a \code{ssn_mod_fit} object.}

\item{proj_years}{A vector of integers specifying the years for which to generate
future predictions. If NULL, it will use all available years from the projection
data. Defaults to NULL.}
}
\value{
A list containing two data tables:
\itemize{
\item \strong{hist}: A data table with historical predictions.
\item \strong{proj}: A data table with future predictions, including
GCM and scenario information.
}
}
\description{
This function takes a fitted SSN model and a list of projection years, then
generates predictions for historical and future conditions.
}
\details{
The function first generates predictions for historical conditions (year 2021)
and then loops through the specified projection years to generate future predictions.
It handles the renaming of columns to match the model's formula and subsets
the data to include only relevant reaches for efficient computation.
}
