% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_data_for_STcon}
\alias{prepare_data_for_STcon}
\title{Prepare data for STcon analysis}
\usage{
prepare_data_for_STcon(in_hydromod_drn, in_net_shp_path)
}
\arguments{
\item{in_hydromod_drn}{A list containing a data.table with hydrological data.}

\item{in_net_shp_path}{A character string specifying the path to the network shapefile.}
}
\value{
A list containing the sites status matrix, network structure matrix,
river distance matrix, and a reference intermittence data.table.
}
\description{
This function takes hydrological and network data, builds a
directed graph of the river network, calculates a river distance matrix,
and formats flow intermittence data into a matrix suitable for STcon analysis.
}
