% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reassign_netids}
\alias{reassign_netids}
\title{Re-assign topologically logical identifiers}
\usage{
reassign_netids(
  rivnet_path,
  strahler_dt,
  in_reaches_hydromod_dt,
  in_reaches_attri_dt,
  outdir,
  country = NULL,
  in_ext = "gpkg"
)
}
\arguments{
\item{rivnet_path}{A character string specifying the path to the input river network shapefile.}

\item{strahler_dt}{A data.table containing the Strahler order for each segment section, typically from the \code{assign_strahler_order} function.}

\item{in_reaches_hydromod_dt}{A data.table with topological data from the hydrological model.}

\item{in_reaches_attri_dt}{A data.table with additional attributes for the reaches from the hydrological model.}

\item{outdir}{A character string specifying the output directory.}

\item{country}{A character string specifying the country. This is used for a series of hard-coded manual corrections for specific countries.}

\item{in_ext}{A character string specifying the file extension for the output file (e.g., 'gpkg').}
}
\value{
A character string with the file path to the newly created and corrected network shapefile.
}
\description{
This function re-assigns a topologically/hydrologically logical 'cat' or 'reach_id' to each segment section in a river network.
The goal is to correct a common problem where these identifiers are incorrectly assigned across confluences in the source data.
The function relies on comparing the spatial network with an independent hydrological model network to make corrections.
}
