% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_ssn_summarized_maps}
\alias{save_ssn_summarized_maps}
\title{Save SSN summary maps}
\usage{
save_ssn_summarized_maps(in_ssn_summarized_maps, out_dir)
}
\arguments{
\item{in_ssn_summarized_maps}{A named list of ggplot objects to be saved.}

\item{out_dir}{The output directory for the saved files.}
}
\value{
A list of file paths to the saved PNG files.
}
\description{
Saves a pre-defined list of key SSN maps to a specified output
directory as high-resolution PNG files.
}
