% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{snap_barrier_sites}
\alias{snap_barrier_sites}
\title{Snap barrier sites to a river network}
\usage{
snap_barrier_sites(
  in_sites_path,
  in_network_path,
  in_sites_idcol,
  attri_to_join,
  out_snapped_sites_path = NULL,
  custom_proj = T,
  overwrite = F
)
}
\arguments{
\item{in_sites_path}{A character string or vector of character strings
specifying the path(s) to the input barrier points file(s).}

\item{in_network_path}{A character string specifying the path to the input network shapefile.}

\item{in_sites_idcol}{A character string specifying the unique ID column for the sites.}

\item{attri_to_join}{A character vector of attribute names from the network
to be joined to the snapped points.}

\item{out_snapped_sites_path}{A character string specifying the output path
for the snapped sites file. If \code{NULL}, a default path is created.}

\item{custom_proj}{A logical value. If \code{TRUE}, a custom projection is created
for distance calculations. Defaults to \code{TRUE}.}

\item{overwrite}{A logical value. If \code{TRUE}, existing files will be overwritten. Defaults to \code{FALSE}.}
}
\value{
A character string with the path to the newly created snapped sites file.
}
\description{
This function takes a vector of barrier points and snaps them to
the nearest point on a corresponding river network.
}
