% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{snap_river_sites}
\alias{snap_river_sites}
\title{Snap river sites to a river network}
\usage{
snap_river_sites(
  in_sites_path,
  in_network_path,
  out_snapped_sites_path = NULL,
  custom_proj = F,
  proj_back = F,
  in_sites_unique_id = "site",
  in_network_unique_id = "UID",
  in_sites_idcol_tomatch = "reach_id",
  in_network_idcol_tomatch = "cat",
  overwrite = F
)
}
\arguments{
\item{in_sites_path}{A character string or vector of character strings
specifying the path(s) to the input site points shapefile(s).}

\item{in_network_path}{A character string specifying the path to the input network shapefile.}

\item{out_snapped_sites_path}{A character string specifying the output path
for the snapped sites file. If \code{NULL}, a default path is created.}

\item{custom_proj}{A logical value. If \code{TRUE}, a custom projection is created
for distance calculations. Defaults to \code{FALSE}.}

\item{proj_back}{A logical value. If \code{TRUE}, the snapped points are re-projected
back to their original projection. Defaults to \code{FALSE}.}

\item{in_sites_unique_id}{A character string specifying the unique ID column
for the sites. Defaults to 'site'.}

\item{in_network_unique_id}{A character string specifying the unique ID column
for the network segments. Defaults to 'UID'.}

\item{in_sites_idcol_tomatch}{A character string specifying the ID column in
the sites data to match with the network. Defaults to 'reach_id'.}

\item{in_network_idcol_tomatch}{A character string specifying the ID column
in the network to match with the sites. Defaults to 'cat'.}

\item{overwrite}{A logical value. If \code{TRUE}, existing files will be overwritten. Defaults to \code{FALSE}.}
}
\value{
A character string with the path to the newly created snapped sites file.
}
\description{
This function takes a vector of site points and snaps them to
the nearest point on a corresponding river network.
It handles projection differences and joins relevant attributes from
the network to the snapped points.
}
