% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{subset_amber}
\alias{subset_amber}
\title{Subset AMBER database and create country-specific files}
\usage{
subset_amber(amber_path, in_hydromod_paths_dt, out_dir, overwrite = T)
}
\arguments{
\item{amber_path}{A character string specifying the path to the full AMBER data file.}

\item{in_hydromod_paths_dt}{A data.table containing paths to hydrological
model data for each country, including catchment boundaries.}

\item{out_dir}{A character string specifying the output directory for the GeoPackage files.}

\item{overwrite}{A logical value. If \code{TRUE}, existing files will be overwritten. Defaults to \code{TRUE}.}
}
\value{
A named list of character strings, where names are countries and
values are the paths to the created AMBER points GeoPackage files.
}
\description{
This function reads the full AMBER database, filters it for
specified countries, and creates a GeoPackage (\code{.gpkg}) file of the barrier
locations for each country. The points are subsetted based on intersection
with country-specific catchment boundaries.
}
