% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{subset_hydrostats}
\alias{subset_hydrostats}
\title{Subset hydrological statistics by date}
\usage{
subset_hydrostats(hydrostats, in_bio_dt)
}
\arguments{
\item{hydrostats}{A list or data.table containing hydrological statistics.
It is expected to have a 'date' column.}

\item{in_bio_dt}{A list of data.tables, where each data.table contains
biological data and a 'date' column.}
}
\value{
The \code{hydrostats} object, filtered to the date range of the biological data.
}
\description{
This function filters a list of hydrological statistics (\code{hydrostats})
to only include data that falls within the date range of biological sampling.
}
