% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize_env}
\alias{summarize_env}
\title{Summarize environmental data}
\usage{
summarize_env(in_env_dt, in_genal_upa)
}
\arguments{
\item{in_env_dt}{A data.table containing raw environmental data with a 'state_of_flow' column.}
}
\value{
A list containing two summarized data tables
(all flows, and only flowing sites) and a ggplot object of the boxplot.
}
\description{
This function takes raw environmental data and computes the mean
value for a set of variables, grouped by drainage basin, site, and stream type.
It also generates a boxplot for visualization.
}
