% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize_network_hydrostats}
\alias{summarize_network_hydrostats}
\title{Summarize network hydrological statistics}
\usage{
summarize_network_hydrostats(
  in_hydromod,
  in_all_date_range,
  in_samp_date_range,
  in_country
)
}
\arguments{
\item{in_hydromod}{A list containing historical hydrological data for different countries.}

\item{in_all_date_range}{A vector of two dates defining the full historical period.}

\item{in_samp_date_range}{A vector of two dates defining the specific sampling period.}

\item{in_country}{A character string specifying the country for which to summarize data.}
}
\value{
A data.table containing summarized hydrological statistics for each river reach.
}
\description{
This function takes hydrological model output and summarizes
basic statistics for a river network across a defined date range.
}
