% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize_sites_hydrocon}
\alias{summarize_sites_hydrocon}
\title{Summarize hydrological and connectivity data}
\usage{
summarize_sites_hydrocon(in_hydrocon_compiled)
}
\arguments{
\item{in_hydrocon_compiled}{A data.table containing compiled hydrological and
connectivity data, typically the output of \code{compile_hydrocon_sites_country}.}
}
\value{
A data.table with a single row per site, containing summarized statistics.
}
\description{
This function computes summary statistics (e.g., mean, max)
for various hydrological and connectivity metrics for each site,
across all dates.
}
