% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{zero_lomf}
\alias{zero_lomf}
\title{Last [non-zero] Observation Moved Forward (lomf)}
\usage{
zero_lomf(x, first = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{first}{(logical) Whether to consider first value as a non-zero value
whose index is moved forward even if it is zero. This prevents having NAs
in the results and somewhat assumes that, for a time series, the day prior
to the first value is non-zero.}
}
\value{
Numeric vector of the indices of the previous non-zero for each
element of the input vector.
}
\description{
Finds the index, for each row, of the previous row with a non-zero value
}
\examples{
test1 <- c(1,1,1,0,0,0,0,1,1)
zero_lomf(test1)
test2 <- c(0,0,0,0,0,1,1,0,1)
zero_lomf(test2, first=FALSE)
zero_lomf(test2, first=TRUE)

}
